backend/requirements.txt
# ============================================================
# Agentic AI BI System — Python Dependencies
# ============================================================

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0

# LLM — Google Gemini
google-generativeai==0.8.0

# Embeddings
sentence-transformers==3.0.0

# Vector store
faiss-cpu==1.8.0

# Utilities
numpy>=1.26.0
python-dotenv==1.0.1
